refactor(cli): extract check/report path into manifest-cli-check (#836)
The grammar-only split still left manifest-cli.ts at 564 lines and
manifest-cli.test.ts at 842 (both still >1.4x the ~400-line guideline),
because most of the original file was check/emit execution, not argv
grammar. Add the third seam: check-report rendering + runCheck (plus the
shared readCommittedManifest/UNEXPECTED_EXIT the emit path also reads) move
into manifest-cli-check.ts, consumed by manifest-cli.ts, which keeps flag
dispatch and the emit/answer-gate path. Tests split the same way, duplicating
setupSandbox/captureStdio per this directory's existing per-file convention.

Pure move, no behavior change; all 46 grammar+execution tests and the full
118-file/1505-test CLI suite stay green. gaia-maintainer binary rebuilt to
match (gaia adopter binary and templates are byte-identical since
manifest-cli.ts is maintainer-only).
